Removing of publisher

+++ Entities/PagedResult.cs
@@ -7,7 +7,7 @@
         public int PageNumber { get; set; }
         public int PageSize { get; set; }
 
-        public int TotalPages => (int)Math.Ceiling((double)TotalRecords / PageSize);
+        public int TotalPages => (int)Math.Ceiling((decimal)TotalRecords / PageSize);
         public bool HasPreviousPage => PageNumber > 1;
         public bool HasNextPage => PageNumber < TotalPages;
     }
