+++ Helper/PaginatedList.cs
@@ -8,7 +8,7 @@
         public PaginatedList(List<T> items, int count, int pageIndex, int pageSize)
         {
             PageIndex = pageIndex;
-            TotalPages = (int)Math.Ceiling(count / (double)pageSize);
+            TotalPages = (int)Math.Ceiling(count / (decimal)pageSize);
 
             this.AddRange(items);
         }
